fix: address Greptile review round 2
- Fix slow relay test: pool.relay() ignores options for existing
  relays (returns cached instance). Renamed to "should succeed when
  one relay has no data" to match actual behavior.
- Fix src/ → e2e/ reverse dependency: move TEST_RELAYS canonical
  definition to src/shared/nostr/test-relays.ts, e2e/helpers/
  re-exports from there.
- Replace literal kind:1 in buildShare with SHORT_TEXT_KIND constant.
- Replace remaining kind:0 literals with METADATA_KIND in tests.
